requirements.txt
azure-storage-blob
pymupdf4llm
tiktoken
azure-cosmos